Update OIDC URLs
The actual OIDC provider URLs in Zuul were a bit muddled.

The OIDC implementation in zuul-web created a second cherrypy
application for oidc, and configured the route map for the first
application to use it, and then also mounted the second application.
The result is that cherrypy was told how to reach the oidc endpoints
twice, once via the route map and once via the application mount.  That
meant that we could previously access the openid configuration document
at both:

`<zuul-root>/oidc/.well-known/openid-configuration`

and

`<zuul-root>/.well-known/openid-configuration`

The original spec indicated that we would support:

* `<zuul-root>/oidc/.well-known/openid-configuration`
* `<zuul-root>/oidc/jwks`

Though it also notes "When a third party service needs to validate such
an ID token it first queries the `.well-known/openid-configuration`
endpoint of the OIDC issuer."

The issuer reported in the openid-configuration document itself is
simply the web root, so the theoretical location of the
openid-configuration document for that issuer would be
`<zuul-root>/.well-known/openid-configuration` (without the "oidc/"
prefix).

Moreover, most users would expect to simply specify <zuul-root> as the
issuer when configuring an OIDC consumer, especially since we didn't
document that they should use an "oidc/" prefix in the URL.

In short, there are two reasons to suppose that the best location for
the openid configuration document should
`<zuul-root>/.well-known/openid-configuration` and that is probably the
location being used by any current users.

As for the second url, that makes perfect sense -- it keeps the jwks
endpoint away from the root, but puts it under "oidc/" instead of "api/"
since it does serve a slightly different purpose.  However, the actual
implementation put it at `/oidc/.well-known/jwks`.  We really shouldn't
put things in well-known that aren't IANA registered.  We can move the
jwks url fairly easily: we just change what's reported in the
openid-configuration document.  But we should support the old URL for a
time so that users can upgrade seamlessly.  We also should make sure to
tell them what the URLs are, since they may need to adjust their proxy
configuration.

Speaking of proxy configuration, this change also updates the
documentation for proxies to include the necessary changes to support
the OIDC endpoints.  It also makes the necessary adjustments to support
whitelabeling, which was supported before, but the URL scheme wasn't as
robust as it could be.  The new scheme mirrors what we do with the api
interface, and so the jwks endpoint for a whitelabel tenant is
`/oidc/tenant/{tenant name}/jwks`.  Including the "tenant" segment
(which was not included before) disambiguates the next segment in the
URL and makes it more future proof.

This change normalizes the URLs to:

`<zuul-root>/.well-known/openid-configuration`

and

`<zuul-root>/oidc/jwks`

and provides some backwards compatibility for earlier jwks urls that
will be removed in a later version.

An alternative would be to lean into the idea of having our OIDC
provider URL (ie, the issuer) be something other than the root URL of
the application.  We could use the two URLs in the spec:

* `<zuul-root>/oidc/.well-known/openid-configuration`
* `<zuul-root>/oidc/jwks`

instead of what is done in this change.  If we do that, we should
change the content of the issuer to `<zuul-root>/oidc`.

One advantage of that is that it would simplify the proxy rules (we
could reduce them to one instead of two).  Or we could stick with the
approach of not prefixing with "oidc/" and then keep the jwks URL under
the well-known directory, so we end up with:

* `<zuul-root>/.well-known/openid-configuration`
* `<zuul-root>/.well-known/jwks`

Which is something that some large providers do, even though it seems
questionable.

Change-Id: Iaed1c34173ea62f2361d1db02452579572adfba9