Launcher: fix loop error in ready node processing
When attempting to assign ready nodes for a request, we can modify
the list of ready nodes over which we are iterating under two
conditions: if we assign one of the ready nodes to our request,
or someone else assigns it to theirs.

If we assign it to our request, we are done iterating over the set
of ready nodes for that label, so we break out of the loop and
start on the next label.

If someone else assigns it, we remove that node from consideration
so we don't attempt to lock it again for the next label, and then
we continue iterating over the loop of nodes.  This can raise an
exception since we're modifying the list while iterating.

To resolve this, coerce our iterator to a list before we begin
iteration.

It is impractical to add a unit test for this since it requires
the node to change state between the start of iteration and our
check to use it.

Change-Id: Iad335a9b61388de2305569b3fac613ff76347913