Launcher: handle resuming accepting a request
If we encounter an error while accepting a request, we may have
already added some nodes to the request but not set it to accepted.
The next time we go through the request loop, we may add the full
set of nodes to the request in addition to the ones already there,
making the request over-full.

To avoid this, skip over any label slots that already have nodes
in the request.

Change-Id: I05b85803d978caafeb97de0222e443b998eadadc