Shutdown log stream socket instead of close
In the streaming log ansible callback, we use the socket close
method to terminate the connection after a command completes.
However, since Ansible forks processes, other processes may
inherit the file descriptor, and the kernel won't actually end
the TCP connection until all of them are closed, which means
they won't actually terminate until the end of the task.

To allow the connections to terminate when we intend, the existing
close() calls are changed to shutdown() which causes immediate
termination of the connection regardless of other processes.
Additionally, we will close the streaming connection before
we create the second connection to delete the file.

This will allow the zuul console server to close its open file
handles which will in turn allow the remote operating system
to finish unlinking the temporary log files sooner.

This sequencing error went unnoticed because on most systems,
it doesn't matter.  But on some versions of Windows (Server 2019
is one), it could cause a problem with a task in a retry loop.
Here is the sequence:

* A win_command task starts writing to a temporary log file on the
  remote server
* zuul_stream requests the file from win_zuul_console
* win_zuul_console opens the file and reads it
* The command completes with an error, but is in a retry loop, so
  the task itself continues
* zuul_stream detects command completion and calls close() on the
  socket and requests the remote side delete the log
* win_zuul_console deletes the log file, but retains the open file
  handle because the connection is still open due to the ineffective
  close() call
* The win_command handler starts the second run of the task, and,
  due to the open file handle, is unable to recreate the deleted
  temporary log file, and returns an error.

The unique behavior in Windows Server 2019 contributes to the problem
in the last step.  Unix and some versions of Windows allow the
creation of the new log file with the same name even while the
old, unlinked file has an open handle.

The updates in this change allow for a strict sequencing where
the file is deleted and the existing open handles closed before
the retried task recreates it.

Change-Id: Ib05380c9a3bfe67606e810271322a0ce91d24691