Handle retries in zuul_stream
To date, we haven't really handled Ansible task retries in
zuul_stream.  When they happen, it works well enough (though see
change Ib05380c9a3bfe67606e810271322a0ce91d24691 for a case
where it didn't) but the behavior is not ideal.  The current
behavior is to output retry information to the executor log, but
in the main job log, we only output the streaming log from the
first attempt of the task, and then we output the result from
the last attempt of the task.  During a long retry period, we
get no other feedback.

This change updates zuul_stream to output the streaming log for
all attempts of a task.

First, a note on the approach: this change causes the remote node
to use a different temporary console log file for each attempt
of the command.  This is because the stream handling in
zuul_stream is implemented in a way that assumes at the
completion of each command, the return code is output into the
log file and that indicates that the command is finished and the
stream should be terminated.  An alternate implementation of
retry handling where we continue to use the same file and append
to it, however, that would require considerable alteration to
zuul_stream.  In such a system, we would need to know at the
completion of the task whether the most recent invocation of the
retried command had finished outputting to the log file, which
means we would also need to add new information to the log file
to indicate the start of a retried command (in order to
differentiate the last attempt from the previous one).  Rather
than re-writing the log streaming handlers, this change uses
unique filenames to achieve the same result with fewer
alterations.

Ansible calls the retry method in the callback after the
completion of an attempt that will be retried.  In our handler,
we will wait for the existing log stream to finish, then start a
new one.  Because retry delays can be long, we won't start
streaming the new one until just before we expect Ansible to
start the next attempt.  This avoids having zuul_stream
repeatedly fail to stream the log.

We output a message to job-output.txt that is similar to the
message that the default callback prints to stdout indicating a
retry is happening.

In order to guarantee a unique filename, we extend our filename
pattern to include another counter.  We already have a counter
for task execution that accounts for multiple "include"s of a
task.  We add a counter for retries of the same task inclusion.
Because retries are still the same execution of a task, we have
no way of coordinating between the callback and the task
executor.  So each of these has an independent copy of the
counter, and we increment it in both.  We presume that nothing
would cause either zuul_stream or the task executor to increment
the counter without the other, and therefore, they will stay in
sync.

A small bug in zuul_stream is corrected: we previously only stored
one instance of a streamer for a particular task, but the
callback may be handling multiple streamers for the same host.
The streamer dictionary is updated to use task and host as a
key.  Even though we expect to handle tasks with the same uuid,
we don't expect that to happen at the same time, therefore we
don't need to add the task counter to the key.  Likewise, we
don't need to add the retry counter since we stop the streamer
when we retry.

The zuul stream tests have been updated to exercise this, and
the Windows versions of these have been run locally.

Change-Id: I6f98ba8d29df90d48ea778f9f945bb9bb343be63