Remove model backwards compat
This removes backwards compatability support for model api versions
before Zuul 11.

This also fixes the tests in test_nodepool which relied on other tests
in the suite initializing opentelemetry.  This caused them to fail when
run in isolation.  To correct that, the test_nodepool tests now initialize
tracing.

One user-facing deprecation is converted to a hard error, so a release
note is included for that.

Change-Id: I2d2e38bf38ea12ecf88cd87239e88828ea0ea5f6