Remove Swagger/OpenAPI components
Current swagger package we use has vulnerabilities and it's not trivial to upgrade it. We need to either complicate existing code and write bunch of unmaintainable overrides, or swap current frontend framework from create-react-app (which had little to no maintenance since 2021 and the React team has officially deprecated it in 02.2025).

Change-Id: If1ce57c8b57e6af046995769b7f8c2a4b9de31b3