Executor: allow configuring termination signal
It looks like when zuul-executor is running in a container (for example
on a kubernetes cluster) and receives a SIGTERM signal (as is the case
when scaling down), the signal is propagated to all processes and
subprocesses within the container. This breaks the graceful stop
workflow, for example when a job just started before the executor
receives the SIGTERM signal, the subprocess call to git to clone the job's
repository will fail with a -15 error.

This patch allows, as a mitigation, to use a different signal to
initiate a graceful stop. To use this solution in a containerized
context, the zuul-executor container must be built with support for
the new signal code.

Change-Id: I1b2b982e5449a0df8ff6fa861d2d630963e8fc5f