Delay timer election until scheduler primed
The timer driver attempts to win the election to emit timer events
as soon as it is started, but until all of the tenants are loaded,
it may not be the best scheduler to do so.  If the cluster is
restarting, then it would be better to let another scheduler handle
events until this one is primed.

There is one downside to this: when starting a system with no
running schedulers, we won't emit events for tenants until all
of them are loaded, even though it would be safe to do so.  Therefore,
we detect that case and go ahead and run the election immediately.

This change also adds some debug messages to help diagnose potential
causes for missing events, as well as expands the exception handling
when we are triggered to catch and log more potential errors.

Change-Id: Ifc73c7206a73a929c43b23f8972e8d19b3f8850c