Fix scheduler openstack memory leak
The OpenStack SDK connection object allocates resources internally
in such a way that it may not be garbage collected unless its
close() method is called.  In the scheduler, we currently create
what is intended to be a throwaway client in order to determine
what image format a provider requires.  Due to this, the scheduler
can leak memory each time it parses image objects.

To avoid this, stop short of creating a full Connection and instead
create only the CloudRegion object, which is all that is needed to
determine the image format.  We will only create Connection objects
when we actually start endpoints in the launcher.

(These will also not be gargabe collected, but that is expected,
because in the launcher, the endpoints are long-lived and currently
have no cleanup process.)

Change-Id: Ie6930dd1752fbeda913971b3e5cc2f6713faa81b