Fix provider selection for image validation
When selecting a provider for a request that explicitly needs a specific
image for validation, we must also check that the upload is valid for
the current provider.

It looks like the original check was accidentally removed in
Ia84f6965d0851eea5a78774d2d8018d0c82d3b6a.

Change-Id: I14f1f3bdb75ff985b24b008e4fc127b1f25ab86a