Allow dynamic addition of images, labels, and flavors
We don't allow dynamic configuration of NIZ objects because we
don't want people to be able to launch resources in cloud environments
without review.  However, we can loosen this restriction so that
images, labels, and flavors can be added dynamically.  We still won't
allow changes to sections or providers, so these object's can't actually
be *used* without being merged.  But it will allow images to be built.

That addresses an issue commonly observed during OpenDev's piloting
of NIZ: that adding a job to build a new image is a two-step process,
whereas it seems like it should be just one.  This will allow a user
to add a new "image" object along with the job that will build that
image in the same change.  This is the main goal of this change;
permitting the same for labels and flavors is done merely for
consistency.

Nonetheless, by adding them, we can actually support defining a new
image and label, and attaching those new objects to a provider
(which may itself be new) all in one change.  The provider won't
actually be used until after it's merged and Zuul reconfigures,
but the syntax will be validated.  The test_label_addition* tests
added in this change explore the limits of how much we can do
in one change.

An unrelated bug related to parsing return data in the executor is
fixed in this change since it exercised the bug.

The WaitForArtifacts method is moved to the base test case to avoid
repitition.

The "nodepool-absent" test fixture is renamed "nodepool" since it
is pretty close to a general-purpose nodepool fixture -- the "absent"
items only come from the tests that use it.

Change-Id: Ib0f2962ec50497f0746a2563d505109b8d61b064