WIP: Don't apply provider defaults upwards
Since we are modeling provider and section configuration on
inheritance, it will likely make more sense to users to have the
"image-defaults", etc, sections apply only to the images, etc,
defined at that inheritance level and below.

In other words, prior to this change, a user could specify
"image-defaults: {iops: 2000}" in a provider, and it would override
"image-defaults: {iops: 1000}" set in the parent section.

That sounds correct, but perhaps less so if an image was attached
at the parent section level.  A user may rightly expect that particular
image to inherit the iops default from that particular section, not
its child provider.

However, there are cases where a user may expect a value set in a
child provider to apply to images attached to parent sections.
The label-defaults of key names or subnets come to mind.  In these
cases, the user wants a consistent configuration at the parent
section level, but to set provider-specific values under that.

Example: https://opendev.org/opendev/zuul-providers/src/branch/master/zuul.d/providers.yaml#L648

Possible solutions:
1) Leave things the way they were before this change.
2) Keep this change, but adjust it so that defaults that are
   added in child providers apply to parents, but only if
   a value hasn't already been set in the parent.

If we go with #1, we should probably keep the part of this change
that ensures unique images, etc, on each provider.  That would allow
users to specifically override values set on parent images.

Option #2 sounds a little confusing, but would probably work in
practice.  Values that should apply more globally would be set
at a higher level, and values that are more localized to regions
would be set at a lower level.  The key is that they must be
omitted at the higher level (otherwise they will attach to
images, etc, defined at that level).

Change-Id: I9813c4537d9cff5586114c3e95d316f79876ca11