Add "final" attribute to image/flavor/label objects
Providers are built from Sections using inheritance.

For convenience, we allow the "image-defaults" and similar
stanzas in child providers and sections to override the values
in their parents, and then apply the end-result to images, etc.

That makes it easy to set labels in a parent section but then
supply the subnet id in a child.

When a section is intended to be shared between tenants, and
images are defined on that section, it may become easy for the
tenant end-users to accidentally override values set on those
images, which would work, but would cause extra uploads of the
images.

To allow operators to prevent this, we introduce a "final"
attribute to image, etc, definitions.

Let's just use "image" as a stand-in for image, label, flavor for
the rest of this commit message.  The same applies to all.

Once the final attribute is set on an image (either directly, or
via image-defaults), all the images definet at that level and
above in the hierarchy will have the current defaults values applied,
and then they will no longer be updated after that.

Children can add more images which may be affected by subsequent
defaults values, but existing ones will not be changed.  (Note:
the "final" attribute itself is subject to inheritance, so any
new images will also be declared final, unless the user switches
it back to "false", which is also valid).

To support a "middle ground" where users want to prevent accidental
modification of parent images, but are okay with deliberate changes
to them, "final" may be set to "allow-override" which will behave
as above, but if a user explicitly redefines an image (with a
duplicate entry in the images list for the same image name), they
may set whatever values they wish on the image.  This allows users
to intentionally create a duplicate image with different values.

If final is set to True and a user attempts to do that, it will
be reported as a configuration error.

Change-Id: I9813c4537d9cff5586114c3e95d316f79876ca11