Fix provider quota calculation
So far we've included the provider resource limits in the same quota
information as the endpoint limits. However, this makes provider
resource limits essentially global endpoint resources limits.

This has the effect that a provider becomes unusable (and pauses) as
soon as the unmanaged quota usage exceeds the provider resource limits.

It looks like we have to distinguish two cases here in order to
determine if a given provider has enough quota:

1. check whether the provider endpoint has in principal enough resources
2. we don't exceed the provider's configured resource limits

For case 1. we need to consider unmanaged quota as well as quota used by
OTHER providers for the same endpoint. The reason for this is, that
there is no 1-1 mapping between an endpoint and a provider.

To cover case 2. we need to use the lower value for a given quota key
between the available endpoint quota (1. above) and the provider
resource limits.

Change-Id: Ifef8c94c64041c17c6fdf900caa0cfbdf8be1264