Run yarn install with --verbose and --frozen-lockfile
This should give us more information on which specific versions of
packages are being installed during the yarn installation process. Using
--frozen-lockfile also makes yarn avoid editing the lockfile in the few
corner cases it may do so and instead return an error if there is a
problem resolving deps with the lockfile.

All of this should give us more control and information about the js
build process.

Change-Id: I970ed0355d00e2beb2f49873710203dd0b370da5