Add --frozen-lockfile to yarn install invocations
There are some corner cases where yarn will do an installation with a
modified lockfile [0]. We don't want that. We want the lockfile to be
the source of truth and error if it isn't possible to install things
that way for us. Add the --frozen-lockfile flag to keep yarn from
editing the lockfile. It should also exit with an error if it wants to
edit things and this flag is set.

[0] https://classic.yarnpkg.com/lang/en/docs/cli/install/#toc-yarn-install

Change-Id: I3385f67e53e3943c96e55ed0b66242569bdb1fa9