Use yarn list instead of yarn install --verbose
Yarn install --verbose is very verbose. While it includes the
information we're interested in, it also includes other information that
is less helpful to the tune of ~82MB for a job-output.txt file when
building Zuul's python packages.

Instead of running yarn install --verbose run yarb list after yarn
install which should give us an accounting of all the packages that are
installed. In theory this will be more efficient in terms of size.

Change-Id: I04710ade79299b855ce858682cce2941185006bb