Add image validation admin method
To allow tenant admins to re-trigger an image validation, this
change adds a REST API endpoint.  A new endpoint is added, rather
than using our existing framework for re-enqueing buildsets, due
to the extra information needed for image validation events.

The web UI is updated in the same change since this is not yet
in production use.  The web interface is modeled on the existing
functionality to delete an image upload.  It appears in the action
menu next to each upload.

Change-Id: Ia08c652058fe4c70e64db50f917039fda874c0e5