Set event state of nodeless requests to completed
We create nodeless nodeset requests with state "fulfilled". However, so
far we did not set the event state to "complete".

This lead to the launcher processing those requests and sending
nodes-provisioned events. In some cases this could also lead to an
NoNodeError bubbling up to the main thread.

Fix this issue by setting the event state to "completed" for nodeless
jobs.

Change-Id: I96131a9bc8b5b0fb216712027e5e25974c94b2a3