WIP: Add subnode support to launcher
This implements support for the concept of "slots" where a single
real node can be used to run multiple jobs.  Since the mechanisms
for assigning and claiming nodes are all based on having distinct
lockable node objects, this is implemented by creating new ZK
node objects for each slot, or subnode.

The main node holds references to the ids of all of its subnodes,
and each subnode holds a refence to the main node.

We do not run a create state machine for the subnodes, instead,
we ignore them until the main node is finalized, then we copy
the relevant fields.  At that point, the subnodes are free to
act idepdentely like normal nodes.

Their delete state machine is a no-op, which results in them
being immediately deleted from ZK.  Once all subnodes are deleted,
the main node may be deleted.

This may be applied to any driver.  This is part of the feature
set of the nodepool "metastatic" driver; once some other features
are added, we will have the metastatic functionality without the
need of a separate provider configuration.

Change-Id: Ib0e657fe24662f04ea8264a22c9f9edebe6a3d8e