Add subnode support to launcher
This implements support for the concept of "slots" where a single
real node can be used to run multiple jobs.  Since the mechanisms
for assigning and claiming nodes are all based on having distinct
lockable node objects, this is implemented by creating new ZK
node objects for each slot, or subnode.

The main node holds references to the ids of all of its subnodes,
and each subnode holds a refence to the main node.

We do not run a create state machine for the subnodes, instead,
we ignore them until the main node is finalized, then we copy
the relevant fields.  At that point, the subnodes are free to
act idepdentely like normal nodes.

Their delete state machine is a no-op, which results in them
being immediately deleted from ZK.  Once all subnodes are deleted,
the main node may be deleted.

This may be applied to any driver.  This is part of the feature
set of the nodepool "metastatic" driver; once some other features
are added, we will have the metastatic functionality without the
need of a separate provider configuration.

This change centralizes node creation in _requestNode: we had two
locations where we created node objects that were nearly identical.
Since this process is getting more complex, centralize it and
account for the differences.  Note: we already missed the fact that
we don't set the initial quota on the requested node when we create
a min-ready node.  In practice, that probably won't come up much,
but it was still technically incorrect and fixing it should lead
to more accurate quota estimates.  There is a race condition when
starting a launcher and creating min-ready nodes; if it attempts the
creation before the quota cache is updated, it will throw an
exception.  To correct this, the quota cache will now insert new
values into the local cache so they can be immediately used rather
than waiting for the ZK round trip to create the cache entries.

Change-Id: Ib0e657fe24662f04ea8264a22c9f9edebe6a3d8e