Pull base python images from quay.io/opendevorg
OpenDev has moved the home for these images back to quay.io/opendevorg.
This means we don't need to pull the docker images that were mirrored to
quay.io/opendevmirror anymore. And in fact we shouldn't as those images
will not get any more updates. Use the new correct location for the
images.

Change-Id: I5944d23f38a2d974c70319e234c245fc809dd5d5