Build container images on Debian Trixie
This updates the base platform for Zuul's container images from Debian
Bookworm to Trixie. The python version remains fixed at python3.11 for
now, but we'll get updated system tools like git.

As part of this change support for very old debuntu versions within the
bindep.txt file is cleaned up to simplify the ruleset.

Trixie does appear to default homedirs to 700 permissions which breaks
the execptations that Zuul's homedir in the container images can be
accessed by other users if running the container as a user other than
Zuul. Address that by updating the permissions on this directory to 755
after it is created by useradd.

Depends-On: https://review.opendev.org/c/opendev/system-config/+/960681
Change-Id: I29ab7879e1e7e84c5abcba20ce8be9acc25bf332