Launcher: don't use the main node with slots
The current slots implementation creates a "main" node as slot 0
and subnodes as the remaining slots.  This change modifies the
process to create subnodes for each slot, including 0, and then
does not use the main node at all.

This simplifies some things (like deciding if a node is actionable),
and makes others more complex (like creating the node and subnodes,
and assigning the subnode to the request).

But the main benefit is that it lets us be more explicit about the
lifecycle of the main and subnodes.  By not treating the main node
as a subnode, we have more choices about what tags we associate with
it and how we want it to appear to the user.  We also reduce the
risk of accidentally relying (perhaps in a job) on something that
only exists in the main node but not on a subnode.

Change-Id: Ie79783a0824f44ce92a01b107097a52e5b253e99