Give quickstart gerritconfig a distinct /var/lib/zuul volume
The gerritconfig container in the quickstart system use the
zuul-executor image to get access to an Ansible installation which is
then used to configure Gerrit. This container requires a volume to be
mounted to /var/lib/zuul, but since gerritconfig doesn't run
zuul-executor we don't actually need that content to by in sync with the
actual executor container. Mount a different distinct volume to the
gerritconfig container so that it can be managed independent of the
executor container's volume.

The motivation here is that in debugging some executor behaviors I
needed to start the executor with a fresh /var/lib/zuul volume. To
delete the volume I first had to delete the gerritconfig container as it
also mounted this volume. This then deleted the gerrit container
automatically for me due to dependencies in the container tree.
Separating the volumes should allow us to make changes to the executor
if necessary without impacting the gerrit install. This should speed up
and simplify debugging in some situations.

Change-Id: Ie05cc3aca0beda20540203be707f97670643db8b