Handle launch failures with subnodes
This ensures we update the state correctly for subnodes in all
places where that should happen.  A test is added to exercise
a launch failure to ensure nodes are cleaned up appropriately.

We short-circuit the checkNode method for subnodes because there
is nothing we need to do there.  We were previously skipping
most of it, but we still ran the nodescan, which is incorrect.
We don't need to do that because we copy data from the main
node after its nodescan completes.

Change-Id: I10c123d870cdc4a2f5981431373d6b24051835a7