Always associate nodes with providers
In earlier NIZ work we elected to allow nodes to "float" between
providers in some circumstances: when a node's request is canceled
before the node is used, or for min-ready nodes.  Floating nodes
can be used by any provider in any tenant with a matching label
configuration.

This is problematic for reused nodes (and nodes with slots) because,
in the case of reuse, a node could be reused in a different tenant,
which means the node will have crossed trust domains (one tenant
could install a trojan horse on the node).  In the case of slots,
two different tenants could be using the same node in different
slots.

If we always associate nodes with providers, we reduce the scope
of this issue by allowing tenant administrators to determine whether
that is acceptable.  If a provider is used in multiple tenants,
then the node may be reused by those tenants.  If a provider appears
only in one tenant, than it will only ever be used by that one
tenant.

Alternative: we could, with some more complexity, only enable
this behavior for nodes with reuse or slots set.  I have started
with this change because it is the simplest to understand and has
the most consistent behavior.

Change-Id: I4b35661195c7d1cb65548d75f6a1c4ad91dd1d32