Launcher: handle reused node failure
In nodepool, after a node is used, if we plan to reuse the node,
we perform a ssh keyscan to verify that we can still reach it.

This change adds similar behavior.  If we can't reach a reusable
node after use, we mark it failed.  If that node is a subnode,
we mark it's main node and any other subnodes we can lock as
failed as well.  This effectively takes the node out of service
ASAP, but any subnodes currently in use will continue through the
normal process.

Change-Id: I954e248e8376f7da05dbab2f29ace4e3c4be4934