Allow base sections with no connection
We have been assuming that every section would be attached to a
connection, and in particular, the only examples in the spec had
the connection declared at the base of the section inheritance
hierarchy (so it applied to that section and all its children and
their providers).

However, it is useful in some circumstances to be able to configure
settings on a base abstract section with no connection attached to
it, and then inherit from those in concrete sections which do have
different connections.  In order to be useful, those connections
will have to be either to the same or similar clouds, but that
is not far-fetched (even opendev could take advantage of this for
some settings).  There doesn't seem to be a reason to insist on
the original behavior, so let's relax this requirement.

This change ensures that we still get a connection for every provider
that we build.  The only requirement is that connection name appear
somewhere in the inheritance hierarchy for that provider.  Once it
does, it may not be changed.  That connection will be used to validate
the settings of every section in the hierarchy (along with the
provider).

Change-Id: Ie0dd8bee2288559560b372c6f8375bbbf190ade1