Launcher: fix endpoint_score
This method is used to determine which launcher should handle
leaked resource cleanup for which endpoints.  In order to correctly
perform the rendezvous hashing necessary, we need to hash a value
that includes the name of the launcher.  Otherwise, since all launchers
will end up with the same scores for each endpoint, it would be
unpredictable which launcher would service that endpoint (it could
be all of them, none of them, or anything in between).  With the
name in the hash, then each launcher-endpoint combination gets
a unique hash, and they will all agree on which launcher will
service each endpoint.

Change-Id: I459a9f88cc256da3ca297abdf306866fd8485af9