Suppress "endpoint other quota" log message
The recent refactor of quota handling added a new log message, but
it should be part of the group of messages that we only emit under
"interesting" circumstances, like if we choose to use a provider
for a node.

Add it to that group so that we don't spam the logs with it.

Change-Id: I1c1e2d2bbe916867942339ddfa2bc3fba7eaa99b