Launcher: clean tempdir at start
If we shutdown while working on image uploads, we can leave
files in the tempdir.  Empty the tempdir at startup to avoid
filling the disk, or leaving files that would conflict with
future re-attempts of uploads.

This adds an extra level, "zuul-launcher", to the temp dir, so
that if users use the default of "/tmp", we will write images
to "/tmp/zuul-launcher/..." and only attempt to delete files
from that directory.  The tests are updated to set the temp
dir under the test root so that the tests use the correct
filesystem and are cleaned up appropriately.

Change-Id: I22bb1baee3ad00132ef7b764f607f9de0c7788fa