Assign unassigned building nodes to requests
We can have unassigned building nodes under a few conditions:

1) min-ready nodes
2) nodes that were originally created for a request but the request
   was deleted
3) subnodes beyond the initial request

In all of these cases, it would be advantageous to assign these
building nodes to new requests ASAP.  In the case of normal nodes,
that will reduce churn and speed response times by matching nodes
that should be ready sooner to requests, rather than starting to
build excess nodes.  In the case of subnodes, we want to pack
real nodes as much as possible before creating another real node
(otherwise we will end up with a lot of sparsely populated real
nodes).

This change treats ready and building nodes the same as far as
assignment to requsets goes.  It prefers ready nodes before building,
and within each category, it orders them by time (in particular,
this should help with packing the real nodes in situations with
subnodes).

Change-Id: I45291bc0f25b4f54d462db575fa3199e32dbd023