Add debug logging for openstack network lookups
OpenDev has found what we think might be a data error with supplying
network information to node boots that results in the cloud complaining
about multiple potential valid networks that prevents not boots from
happening. It isn't clear where this bad information might be sneaking
in so add a bit more debugging info to the _findNetworks() lookup
function. This shouldn't spam the logs because we use lru caches to
cache the result which means this method should be called very
infrequently.

While we are at it we fix an adjacent Exception's message generation.
OpenStackEndpoints don't have provider values. But we do have a
connection with a name and a region attribute that we can use to provide
the relevant info. We don't need to explicitly provide this when logging
the found network details as the logger itself contains that
information.

Change-Id: Ib3c7623e954de1849ea3c89a01498ebd12d5d923