Do not set is_template for regex project stanzas
In change I2958a9bb01ca5261604773de742b6b652fca4cae we added an
"is_template" flag on the ProjectConfig object so that we could
distinguish between project stanzas and the templates which are
referenced by them.  Unfortunately, during review, we did not
notice that there were two locations where this flag needed to
be set to False, and we only set it in one location.  That
location is on the code path where we are parsing a non-regex
project stanza.

This mistake meant that if a project had only regex-based stanzas,
then they would not appear in the UI, or may display the wrong
set of jobs.

The code has since been refactored, but that conditional remains,
so that we only set the flag to False for non-regex project stanzas.

This should be set for regex project stanzas as well, since the
intent is for the web ui to merge the templates specified by any
given project stanza into the project stanza itself.  This should
be true for regular and regex based stanzas.  This change corrects
that.

Since that change, the is_template field has grown another use:
it is used to determine what branch matcher to use for a project
stanza (in ProjectConfig.getBranchMatcher()).  Fortunately, fixing
this bug does not produce a behavior change.  The structure of
that method is:

  if not self.is_template:
    if same_project:

The old behavior would never enter the first conditional on a
regex project because we were erroneously setting is_template to
True.  When we correct this, we will now enter the conditional on
all project stanzas, including regex projects.

The second conditional is still always false for regex projects.
That's because the same_project check is a direct string
comparison (we do not evaluate the regex).  The result is the
same behavior: we pass over action in that section and honor the
pragmas and implied branch matchers.

This is the correct behavior, because the intent is to define
intuitive behavior for a project stanza with no name -- that
should always have a branch matcher for the branch where it is
defined (so that the user sees a project stanza in a file and it
means "this project, this branch").  A project stanza with no
name gets the fully qualified name of the project, and that's
what we check with the same_project check.  (And it's good that
the behavior is the same if the user explicitly specifies the full
name.)

Change-Id: I5e1549451cd900661ac3a9e7ba27360d9450b587