Use a subnode for request assignment
This allows us to update the request that a node is assigned to
while the node is building.  We can already do that for subnodes,
but this allows us to do that for real nodes.  This will allow
better utilization of nodes whose requests are canceled while
building (eg, gate reset or new patchset).

Change-Id: I894af04db659f8a0d92f6d57bacb052c5728e9a8