Allow section inheritance between projects
In I570028d8275734ad1e0e8aa776f227ad2f41da8e we disallowed a section
from inheriting from a section in a different project.  The simplified
reasoning is that while we can uniquely identify sections, if section-one
inherits from section-two, since the reference is made with a
non-fully-qualified name, section-two might be different in different
tenants.  If we store any section state information in ZooKeeper, then
a given single provider may either reference two different sections
in two different tenants, or which one it should reference might be
ambiguous.

However, we are not currently storing any section information in ZK,
and all of our configuration information is flattened when creating
the providers, which means that we will resolve the different sections
appropriately when flattening.

Since providers already reference their parent sections by name and
those sections can be different in different tenants, all of the issues
that might be introduced by this change are already present and
addressed by that mechanism.

The only potential loss in functionality is the idea of sections as
first-class ZK objects with their own quota values.  We do not anticipate
supporting that kind of quota handling (instead, we have endpoint quotas
and provider quotas -- not section quotas).  If we did want to add
section quotas in the future, it may be more difficult -- we will
need to deal with the possible multiple resolution of section inheritance
in that case.

Meanwhile, allowing inheritance across projects will give operators
more flexibility to build up configuration using a combination of
central and distributed config, as is typical in Zuul.  This is especially
useful since we do not allow setting flavors on providers.  With this
change, operators can define some flavors in a central repo section, and
then add tenant-specific flavors in another section.

Change-Id: I1d3f91aad7a282bd250c01e35a73c11c00224878